A수정

+++ app/src/main/java/cosmic/com/pkwprj/model/GithubOwner.kt
@@ -5,11 +5,11 @@ class GithubOwner {
     lateinit var login: String
     lateinit var avatar_url: String
     lateinit var html_url: String
-    var score: Int = 0
+    var score:Float = 0.0f
 
     constructor() {}
 
-    constructor(login: String, avatar_url: String, html_url: String, score: Int) {
+    constructor(login: String, avatar_url: String, html_url: String, score: Float) {
         this.login = login
         this.avatar_url = avatar_url
         this.html_url = html_url
